Exclude linked children from Public Components
fixes [OSF-7612]

Also turns defunct APIv1 views tests into serializer tests
